slow saving, metadata yaml, epix communication
Got slow time series back up and running-- and button stays checked
while time series is collecting, then unchecks. Fixed string formatting
so that the metadata yaml saves nicely like it used to. The fast time
series and forceful over-writing checks are the things left to do to get
this back to where it was before it was refactored... plus its new
features!